Set up event handlers for Example and Peaceful Pummeling that catch a subset of events. Others are handled by the Common Event Handler
