+++ tests/res/event/event_purifier.json
